Moving the VkDescriptorSet out of the signature and into the primitive
