+++ sources/Graphics/Vulkan/Graphics/VulkanGraphicsPipelineSignature.cs
@@ -3,7 +3,6 @@
 using System;
 using TerraFX.Interop.Vulkan;
 using TerraFX.Threading;
-using static TerraFX.Interop.Vulkan.VkDescriptorPoolCreateFlags;
 using static TerraFX.Interop.Vulkan.VkDescriptorType;
 using static TerraFX.Interop.Vulkan.VkObjectType;
 using static TerraFX.Interop.Vulkan.VkShaderStageFlags;
@@ -19,8 +18,6 @@ namespace TerraFX.Graphics;
 /// <inheritdoc />
 public sealed unsafe class VulkanGraphicsPipelineSignature : GraphicsPipelineSignature
 {
-    private readonly VkDescriptorPool _vkDescriptorPool;
-    private readonly VkDescriptorSet _vkDescriptorSet;
     private readonly VkDescriptorSetLayout _vkDescriptorSetLayout;
     private readonly VkPipelineLayout _vkPipelineLayout;
 
@@ -30,129 +27,13 @@ public sealed unsafe class VulkanGraphicsPipelineSignature : GraphicsPipelineSig
     internal VulkanGraphicsPipelineSignature(VulkanGraphicsDevice device, ReadOnlySpan<GraphicsPipelineInput> inputs, ReadOnlySpan<GraphicsPipelineResource> resources)
         : base(device, inputs, resources)
     {
-        var vkDescriptorPool = CreateVkDescriptorPool(device, resources);
-        _vkDescriptorPool = vkDescriptorPool;
-
         var vkDescriptorSetLayout = CreateVkDescriptorSetLayout(device, resources);
         _vkDescriptorSetLayout = vkDescriptorSetLayout;
 
-        _vkDescriptorSet = CreateVkDescriptorSet(device, vkDescriptorPool, vkDescriptorSetLayout);
         _vkPipelineLayout = CreateVkPipelineLayout(device, vkDescriptorSetLayout);
 
         _ = _state.Transition(to: Initialized);
         Name = nameof(VulkanGraphicsPipelineSignature);
-
-        static VkDescriptorPool CreateVkDescriptorPool(VulkanGraphicsDevice device, ReadOnlySpan<GraphicsPipelineResource> resources)
-        {
-            var vkDescriptorPoolSizes = UnmanagedArray<VkDescriptorPoolSize>.Empty;
-
-            try
-            {
-                // We split this into two methods so the JIT can still optimize the "core" part
-                return CreateVkDescriptorPoolInternal(device, resources, ref vkDescriptorPoolSizes);
-            }
-            finally
-            {
-                vkDescriptorPoolSizes.Dispose();
-            }
-        }
-
-        static VkDescriptorPool CreateVkDescriptorPoolInternal(VulkanGraphicsDevice device, ReadOnlySpan<GraphicsPipelineResource> resources, ref UnmanagedArray< VkDescriptorPoolSize> vkDescriptorPoolSizes)
-        {
-            var vkDescriptorPool = VkDescriptorPool.NULL;
-
-            if (resources.Length != 0)
-            {
-                var vkDescriptorPoolCreateInfo = new VkDescriptorPoolCreateInfo {
-                    sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO,
-                    flags = VK_DESCRIPTOR_POOL_CREATE_FREE_DESCRIPTOR_SET_BIT,
-                    maxSets = 1,
-                };
-
-                var vkDescriptorPoolSizesCount = 0u;
-                var constantBufferCount = 0u;
-                var textureCount = 0u;
-
-                for (var resourceIndex = 0; resourceIndex < resources.Length; resourceIndex++)
-                {
-                    var resource = resources[resourceIndex];
-
-                    switch (resource.Kind)
-                    {
-                        case GraphicsPipelineResourceKind.ConstantBuffer:
-                        {
-                            if (constantBufferCount == 0)
-                            {
-                                vkDescriptorPoolSizesCount++;
-                            }
-                            constantBufferCount++;
-                            break;
-                        }
-
-                        case GraphicsPipelineResourceKind.Texture:
-                        {
-                            if (textureCount == 0)
-                            {
-                                vkDescriptorPoolSizesCount++;
-                            }
-                            textureCount++;
-                            break;
-                        }
-
-                        default:
-                        {
-                            break;
-                        }
-                    }
-                }
-
-                vkDescriptorPoolSizes = new UnmanagedArray<VkDescriptorPoolSize>(vkDescriptorPoolSizesCount);
-                var vkDescriptorPoolSizesIndex = 0u;
-
-                if (constantBufferCount != 0)
-                {
-                    vkDescriptorPoolSizes[vkDescriptorPoolSizesIndex] = new VkDescriptorPoolSize {
-                        type = VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER,
-                        descriptorCount = constantBufferCount,
-                    };
-                    vkDescriptorPoolSizesIndex++;
-                }
-
-                if (textureCount != 0)
-                {
-                    vkDescriptorPoolSizes[vkDescriptorPoolSizesIndex] = new VkDescriptorPoolSize {
-                        type = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER,
-                        descriptorCount = textureCount,
-                    };
-                    vkDescriptorPoolSizesIndex++;
-                }
-
-                vkDescriptorPoolCreateInfo.poolSizeCount = (uint)vkDescriptorPoolSizes.Length;
-                vkDescriptorPoolCreateInfo.pPoolSizes = vkDescriptorPoolSizes.GetPointerUnsafe(0);
-
-                ThrowExternalExceptionIfNotSuccess(vkCreateDescriptorPool(device.VkDevice, &vkDescriptorPoolCreateInfo, pAllocator: null, &vkDescriptorPool));
-            }
-
-            return vkDescriptorPool;
-        }
-
-        static VkDescriptorSet CreateVkDescriptorSet(VulkanGraphicsDevice device, VkDescriptorPool vkDescriptorPool, VkDescriptorSetLayout vkDescriptorSetLayout)
-        {
-            var vkDescriptorSet = VkDescriptorSet.NULL;
-
-            if (vkDescriptorPool != VkDescriptorPool.NULL)
-            {
-                var descriptorSetAllocateInfo = new VkDescriptorSetAllocateInfo {
-                    sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_ALLOCATE_INFO,
-                    descriptorPool = vkDescriptorPool,
-                    descriptorSetCount = 1,
-                    pSetLayouts = &vkDescriptorSetLayout,
-                };
-                ThrowExternalExceptionIfNotSuccess(vkAllocateDescriptorSets(device.VkDevice, &descriptorSetAllocateInfo, &vkDescriptorSet));
-            }
-
-            return vkDescriptorSet;
-        }
 
         static VkDescriptorSetLayout CreateVkDescriptorSetLayout(VulkanGraphicsDevice device, ReadOnlySpan<GraphicsPipelineResource> resources)
         {
@@ -294,35 +175,13 @@ public sealed unsafe class VulkanGraphicsPipelineSignature : GraphicsPipelineSig
 
         set
         {
-            _name = Device.UpdateName(VK_OBJECT_TYPE_DESCRIPTOR_POOL, VkDescriptorPool, value);
-            _ = Device.UpdateName(VK_OBJECT_TYPE_DESCRIPTOR_SET, VkDescriptorSet, value);
+            _name = Device.UpdateName(VK_OBJECT_TYPE_PIPELINE_LAYOUT, VkPipelineLayout, value);
             _ = Device.UpdateName(VK_OBJECT_TYPE_DESCRIPTOR_SET_LAYOUT, VkDescriptorSetLayout, value);
-            _ = Device.UpdateName(VK_OBJECT_TYPE_PIPELINE_LAYOUT, VkPipelineLayout, value);
         }
     }
 
     /// <inheritdoc cref="GraphicsDeviceObject.Service" />
     public new VulkanGraphicsService Service => base.Service.As<VulkanGraphicsService>();
-
-    /// <summary>Gets the <see cref="Interop.Vulkan.VkDescriptorPool" /> for the pipeline.</summary>
-    public VkDescriptorPool VkDescriptorPool
-    {
-        get
-        {
-            AssertNotDisposedOrDisposing(_state);
-            return _vkDescriptorPool;
-        }
-    }
-
-    /// <summary>Gets the <see cref="Interop.Vulkan.VkDescriptorSet" /> for the pipeline.</summary>
-    public VkDescriptorSet VkDescriptorSet
-    {
-        get
-        {
-            AssertNotDisposedOrDisposing(_state);
-            return _vkDescriptorSet;
-        }
-    }
 
     /// <summary>Gets the underlying <see cref="Interop.Vulkan.VkDescriptorSetLayout" /> for the pipeline.</summary>
     public VkDescriptorSetLayout VkDescriptorSetLayout
@@ -352,31 +211,12 @@ public sealed unsafe class VulkanGraphicsPipelineSignature : GraphicsPipelineSig
         if (priorState < Disposing)
         {
             var vkDevice = Device.VkDevice;
-            var vkDescriptorPool = _vkDescriptorPool;
 
-            DisposeVkDescriptorSet(vkDevice, vkDescriptorPool, _vkDescriptorSet);
             DisposeVkDescriptorSetLayout(vkDevice, _vkDescriptorSetLayout);
-            DisposeVkDescriptorPool(vkDevice, _vkDescriptorPool);
             DisposeVkPipelineLayout(vkDevice, _vkPipelineLayout);
         }
 
         _state.EndDispose();
-
-        static void DisposeVkDescriptorPool(VkDevice vkDevice, VkDescriptorPool vkDescriptorPool)
-        {
-            if (vkDescriptorPool != VkDescriptorPool.NULL)
-            {
-                vkDestroyDescriptorPool(vkDevice, vkDescriptorPool, pAllocator: null);
-            }
-        }
-
-        static void DisposeVkDescriptorSet(VkDevice vkDevice, VkDescriptorPool vkDescriptorPool, VkDescriptorSet vkDescriptorSet)
-        {
-            if (vkDescriptorSet != VkDescriptorSet.NULL)
-            {
-                _ = vkFreeDescriptorSets(vkDevice, vkDescriptorPool, 1, &vkDescriptorSet);
-            }
-        }
 
         static void DisposeVkDescriptorSetLayout(VkDevice vkDevice, VkDescriptorSetLayout vulkanDescriptorSetLayout)
         {
